task deleted by task owner

+++ src/controller/task.controller.js
@@ -166,20 +166,40 @@ export const updateTask = async (req, res) => {
 export const deleteTask = async (req, res) => {
   try {
     const taskId = req.params.id;
+    console.log(taskId);
+    const userId = req.user.id;
+    const task = await prisma.task.findUnique({
+      where: {
+        id: taskId,
+      },
+    });
+    console.log(task);
+    if (!task) {
+      return res.status(404).json({
+        sucess: false,
+        message: "task not found",
+      });
+    }
+    if (task.userId !== userId) {
+      return res.status(403).json({
+        sucess: false,
+        message: "you are not allowed to delete task",
+      });
+    }
     const deleteTask = await prisma.task.delete({
       where: {
         id: taskId,
       },
     });
-    res.status(200).json({
+    return res.status(200).json({
       sucess: true,
       message: "Task Deleted SuccessFully",
       deleteTask,
     });
   } catch (error) {
-    res.status(500).json({
+    return res.status(500).json({
       sucess: false,
-      message: "internal server errror",
+      message: "internal sererrrrver errror",
     });
   }
 };
